Refactored cl_ros2_timer components to Header only
