+++ smacc2_client_library/cl_ros2_timer/CMakeLists.txt
@@ -23,7 +23,6 @@ include_directories(
 
 add_library(${PROJECT_NAME}
   src/cl_ros2_timer.cpp
-  src/components/cp_timer_listener1.cpp
 )
 target_link_libraries(${PROJECT_NAME} SHARED ${Boost_LIBRARIES} smacc2)
 ament_target_dependencies(${PROJECT_NAME} smacc2)
